fix get articles

+++ app/src/main/java/com/dicoding/emodiary/data/remote/response/ArticleItem.kt
@@ -68,5 +68,5 @@ data class DetectedExtensions(
     val suara: Int? = null,
 
     @field:SerializedName("skor")
-    val skor: Int? = null
+    val skor: Float? = null
 ) : Parcelable
